interface changes
1)adjusted main activity to accomodate buttons. overrides continue not
to work. not sure how to solve the problem.
2)created a sub-page that allows you to compose a message. no buttons on
it work.
3)buttons on the primary interface have been made. they do not work, but
button listeners have been added to the main activity in an attempt to
get them to work.
COMPOSE PAGE -> can go from main activity page to this page to compose a
message.
DISPLAYMESSAGEACTIVITY -> supposedly you can use this activity to send
your messages. however, it does not work.