compose page view messages
added a list veiw to the compose page to accomodate the veiwing of
outgoing and incoming messages (like an average texting app)

runs but only when attributes of the list veiw are commented out. list
veiw needs to be altered in a way that allows it to be dynamically
updated. still working on it.

uploaded the content for the sake of team clarity.